Support schema generation hooks
This change is an attempt to address pydantic/pydantic#9237 and pydantic/pydantic#8279, both of which request the ability to make modifications to the schema generated for fields nested within third-party types which cannot be controlled.

For example, given some type coming from a third-party codebase:

```python
@dataclass
class ThirdPartyType:
    dataframe: pd.DataFrame
```

... a user may still want to be able to determine how the schema for `dataframe` is generated.

This change introduces the concept of "schema generation hooks", which are functions that can be registered with a `TypeAdapter` to allow for schema generation to be customised dynamically. A schema generation hook takes in some `obj` (the type annotation for which the `GenerateSchema` will generate a schema) and some `handler` callback (the next hook in the chain, which itself takes `obj` and returns a `CoreSchema`), and returns a `CoreSchema`. The hook's implementer can choose whether to call the handler or stop the chain short.

The final function in the chain will always be the standard `GenerateSchema` class' generation function, in order to allow for a fallback.

The rationale behind using `TypeAdapter` for this functionality is that it already exposes an existing, supported interface to generating core schemas for arbitrary types. Other considered approaches were:

1. Define a new interface to generating core schemas;
2. Add support in model configuration to be applied recursively.

I was wary of (1) because of the sheer size of the change, and the amount of extra stuff it could make public which isn't currently. I attempted (2), but ultimately it was tricky to get it to work with more deeply-nested models because there's no current concept of "ancestor" models in a dependency tree affecting the configuration for "descendent" models.

*Example usage*

If we take the third-party type defined above, and add some models which depend on it:

```python
from dataclasses import dataclass
import pandas as pd
from pydantic import TypeAdapter, ConfigDict, with_config
from typing import Any

@dataclass
class ThirdPartyType:
    dataframe: pd.DataFrame

@dataclass
class MyNestedModel:
    third_party_type: ThirdPartyType

@dataclass
@with_config(ConfigDict(arbitrary_types_allowed=True))
class MyTopLevelModel:
    nested_model: MyNestedModel
```

... We can define some hook to modify the `DataFrame`'s schema, and pass that into `TypeAdapter`:

```python
from pydantic_core import CoreSchema, core_schema
from pydantic.hooks import WrappedCoreSchemaGenerator

def my_dataframe_schema_hook(obj: Any, next: WrappedCoreSchemaGenerator) -> CoreSchema:
    if obj == pd.DataFrame:
        return core_schema.no_info_after_validator_function(
            function=pd.DataFrame.from_dict,
            schema=core_schema.dict_schema(
                keys_schema=core_schema.str_schema(),
                values_schema=core_schema.list_schema(items_schema=core_schema.int_schema())
            )
        )
    return next(obj)

adapter = TypeAdapter(MyTopLevelModel, schema_generation_hooks=[
    my_dataframe_schema_hook
])
```

We can then test if our validation function has been applied correctly:

```python
adapter.validate_python({
    'nested_model': {
        'third_party_type': {
            'dataframe': {'a': [1], 'b': [2]}
        }
    }
})
```

This will create a `MyTopLevelModel` instance, and nested all the way down in the `dataframe` field in `ThirdPartyType` will be a dataframe:

```
   a  b
0  1  2
```